image_to_text/requirements.txt
streamlit==1.31.0
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.2.0
python-docx==1.0.1
fpdf==1.7.2 